@@ -1,16 +1,36 @@
 module github.com/IBM/hpdb-go-sdk
 
-go 1.16
+go 1.20
 
 require (
-	github.com/IBM/go-sdk-core/v5 v5.10.2
-	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d // indirect
-	github.com/go-openapi/errors v0.20.3 // indirect
-	github.com/go-openapi/strfmt v0.21.3
-	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/IBM/go-sdk-core/v5 v5.13.2
+	github.com/go-openapi/strfmt v0.21.5
 	github.com/onsi/ginkgo v1.16.5
-	github.com/onsi/gomega v1.18.0
-	github.com/stretchr/testify v1.8.0
-	go.mongodb.org/mongo-driver v1.10.2 // indirect
+	github.com/onsi/gomega v1.27.6
+	github.com/stretchr/testify v1.8.2
+)
+
+require (
+	github.com/asaskevich/govalidator v0.0.0-20230301143203-a9d515a09cc2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/fsnotify/fsnotify v1.6.0 // indirect
+	github.com/go-openapi/errors v0.20.3 // indirect
+	github.com/go-playground/locales v0.14.1 // indirect
+	github.com/go-playground/universal-translator v0.18.1 // indirect
+	github.com/go-playground/validator/v10 v10.13.0 // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
+	github.com/hashicorp/go-retryablehttp v0.7.2 // indirect
+	github.com/leodido/go-urn v1.2.3 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/nxadm/tail v1.4.8 // indirect
+	github.com/oklog/ulid v1.3.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.mongodb.org/mongo-driver v1.11.3 // indirect
+	golang.org/x/crypto v0.7.0 // indirect
+	golang.org/x/net v0.8.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/text v0.8.0 // indirect
+	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
